Consolidate addons
