--- demo-addon/.npmignore
@@ -1,21 +1,0 @@
-/bower_components
-/config/ember-try.js
-/dist
-/tests
-/tmp
-**/.gitkeep
-.bowerrc
-.editorconfig
-.ember-cli
-.eslintrc.js
-.gitignore
-.watchmanconfig
-.travis.yml
-bower.json
-ember-cli-build.js
-testem.js
-
-# ember-try
-.node_modules.ember-try/
-bower.json.ember-try
-package.json.ember-try
