--- demo-addon/.watchmanconfig
@@ -1,3 +1,0 @@
-{
-  "ignore_dirs": ["tmp", "dist"]
-}
