--- demo-addon/tests/dummy/app/resolver.js
@@ -1,3 +1,0 @@
-import Resolver from 'ember-resolver';
-
-export default Resolver;
