--- demo-addon/tests/dummy/app/router.js
@@ -1,12 +1,0 @@
-import EmberRouter from '@ember/routing/router';
-import config from './config/environment';
-
-const Router = EmberRouter.extend({
-  location: config.locationType,
-  rootURL: config.rootURL
-});
-
-Router.map(function() {
-});
-
-export default Router;
