--- demo-addon/tests/dummy/app/templates/application.hbs
@@ -1,3 +1,0 @@
-<h2 id="title">Welcome to Ember</h2>
-
-{{outlet}}
