+++ demo-app/app/styles/app.scss
@@ -1,3 +1,3 @@
 // We import our addons styles into the application
 
-@import 'demo-addon';
+@import 'demo-ui-addon';
