+++ demo-app/app/templates/application.hbs
@@ -7,8 +7,8 @@
   <nav>
     <ul>
       <li>{{#link-to "demo-engine.index"}}Engine{{/link-to}}</li>
-      <li><a href="" {{action 'setThemeTwo'}}>Change to Theme 2</a></li>
       <li><a href="" {{action 'setThemeOne'}}>Change to Theme 1</a></li>
+      <li><a href="" {{action 'setThemeTwo'}}>Change to Theme 2</a></li>
     </ul>
   </nav>
 </header>
