+++ demo-app/app/templates/index.hbs
@@ -1,5 +1,5 @@
 <img src="assets/ember-construction.png" alt="logo">
 
 {{#themed-button}}
-  Theme 1
+  {{theme}}
 {{/themed-button}}
