+++ demo-engine/addon/templates/application.hbs
@@ -1,5 +1,5 @@
 <img src="engines-dist/demo-engine/assets/images/ember-learn.png" alt="logo">
 
 {{#themed-button}}
-  Theme 2
+  {{theme}}
 {{/themed-button}}
