+++ demo-engine/tests/dummy/app/templates/application.hbs
@@ -3,8 +3,8 @@
 <header>
   <nav>
     <ul>
-      <li><a href="" {{action 'setThemeTwo'}}>Change to Theme 2</a></li>    {{!-- {{else}} --}}
       <li><a href="" {{action 'setThemeOne'}}>Change to Theme 1</a></li>
+      <li><a href="" {{action 'setThemeTwo'}}>Change to Theme 2</a></li>
     </ul>
   </nav>
 </header>
