+++ demo-engine/tests/dummy/app/templates/index.hbs
@@ -1,1 +1,5 @@
 <img src="engines-dist/demo-engine/assets/images/ember-learn.png" alt="logo">
+
+{{#themed-button}}
+  {{theme}}
+{{/themed-button}}
