+++ demo-ui-addon/app/services/asset-map.js
