+++ demo-ui-addon/app/services/theme-changer.js
