+++ demo-ui-addon/app/styles/demo-ui-addon.scss
@@ -2,7 +2,7 @@
   This is your addon's base style that will be included into
   your host app's `app.scss` with:
 
-  @import 'demo-addon';
+  @import 'demo-ui-addon';
 */
 
 body {
