+++ demo-ui-addon/app/styles/theme-2.scss
