+++ demo-ui-addon/tests/dummy/app/templates/application.hbs
@@ -1,7 +1,3 @@
 <h2 id="title">Welcome to Ember</h2>
 
 {{outlet}}
-
-{{#themed-button}}
-  Button
-{{/themed-button}}
